+++ Changelog.md
@@ -29,3 +29,5 @@ All notable changes to this project will be documented in this file.
 - WKBWriter adds 4 null bytes? (#475)
 - Exception when calling feature.Intersects(otherFeature) (#746)
 - Remove Map Frame context menu doesn't work (#237)
+- ReprojectPoints() not working for more than 1 point for Stereographic transforms. (#781)
+